+++ packages/connectors/hull-warehouse/server/glue.js
@@ -134,10 +134,14 @@ const glue = {
         postgresJdbc("closeDatabaseConnectionIfExists"),
         route("accountSchemaUpdateStart"),
         route("userSchemaUpdateStart"),
-        cacheSet({ key: "databaseSettings", ttl: 99999999 }, "${currentDatabaseSettings}")
+        cacheSet({ key: "databaseSettings", ttl: 3600 }, "${currentDatabaseSettings}")
       ]),
       elif: [
         ifL([
+          // only check if "containsNewAttribute" if we're sending everything
+          // otherwise, it wouldn't make sense to check because if we were selecting the attributes we know it's in the schema
+          // and we would have gotten a ship:update if the outgoing attributes were changed, so would have resync'd anyway
+            cond("isEqual", "${connector.private_settings.send_all_user_attributes}", true),
             cond("notEmpty", input("[0].user")),
             postgresJdbc("containsNewAttribute", {
               messages: obj(input()),
@@ -148,6 +152,7 @@ const glue = {
           lockL("${connector.id}", route("userSchemaUpdateStart"))
         ),
         ifL([
+            cond("isEqual", "${connector.private_settings.send_all_user_attributes}", true),
             cond("notEmpty", input("[0].account")),
             postgresJdbc("containsNewAttribute", {
               messages: obj(input()),